requirements.txt
# Sage Voice Assistant Dependencies

# LiveKit Agents Framework
livekit-agents[silero]>=1.3.0
livekit-plugins-openai>=1.0.0

# Environment and HTTP
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Admin Panel
fastapi>=0.109.0
uvicorn>=0.27.0
